rename test file to match mode
