+++ reward_index/tests/integrations/main.rs
@@ -1,5 +1,5 @@
 mod common;
 
-mod escrowed_mobile;
 mod iot;
 mod mobile;
+mod mobile_escrowed;
